Solving for python 2.7 where round does not return int

+++ prime_numbers.py
@@ -5,7 +5,7 @@ def is_prime(x):
     if x <= 1:
         return False
 
-    for i in range(2, round(sqrt(x)) + 1):
+    for i in range(2, int(round(sqrt(x)) + 1)):
         if x % i == 0:
             return False
